removed composite widget
